+++ tests/ecmascript/test-regexp-charclass-canon-individual.js
@@ -1,0 +1,35 @@
+/*
+ *  Stress test for case insensitive regexp normalization, all BMP
+ *  codepoints.
+ *
+ *  This is disabled by default because the test takes hours to run.
+ *  Run manually when changing regexp code.
+ */
+
+/*@include util-regexp.js@*/
+
+/*---
+{
+    "skip": true
+}
+---*/
+
+/*===
+===*/
+
+function test() {
+    var re;
+    var i;
+
+    for (i = 0x0; i <= 0xffff; i++) {
+        re = RegExpUtil.makeCaseInsensitiveCharClassRegExp(i, i);
+        print(i, re.source, RegExpUtil.getRegExpSingleCharMatches(re));
+    }
+
+}
+
+try {
+    test();
+} catch (e) {
+    print(e.stack || e);
+}